refactor: salary 은행 통신 step 과정 통합
